+++ server/tests-py/test_graphql_queries.py
@@ -210,6 +210,9 @@ class TestGraphQLQueryBasicBigquery:
     def test_basic_remote_join(self, hge_ctx, transport):
         check_query_f(hge_ctx, self.dir() + "/basic_remote_joins.yaml", transport)
 
+    def test_nested_array_relationships(self, hge_ctx, transport):
+        check_query_f(hge_ctx, self.dir() + "/nested_array_relationships.yaml", transport)
+
     @classmethod
     def dir(cls):
         return 'queries/graphql_query/bigquery'
@@ -1240,7 +1243,7 @@ class TestMissingUnauthorizedRoleAndCookie:
         check_query_f(hge_ctx, self.dir() + '/cookie_header_absent_unauth_role_not_set.yaml', transport, add_auth=False)
 
 @usefixtures('per_class_tests_db_state')
-class TestGraphQLExplain:
+class TestGraphQLExplainCommon:
     @classmethod
     def dir(cls):
         return 'queries/explain'